refactor: mve invites components
