+++ apps/web/src/pages/invites/components/MemberRole.tsx
@@ -1,7 +1,7 @@
 import { MemberRoleEnum } from '@novu/shared';
 import styled from '@emotion/styled';
 
-import { Dropdown, Tag } from '../../design-system';
+import { Dropdown, Tag } from '../../../design-system';
 
 export function MemberRole({ member, onChangeMemberRole, isEnableMemberActions }) {
   const TagElement = () => {
